CommandCatalog no longer populates itself with default commands
This was screwing up the command container state for tests. It's also better this way in the long run, since the catalog should contain--not discover--its commands (especially in a plugin model).

Currently it's being seeded on app startup. May not be the final place, but it's a better place (so far) than the catalog itself.