MainViewModel now defers preview logic to a CommandParser instance
This doesn't need to be an interface since it doesn't have any new dependencies; it needs the CommandCatalog, but we already have an instance of that. This feels safe to let the view model create a real parser while passing it the catalog interface it already has.

It also creates the parser instance in the constructor so it can retain it, since it'll be called a lot (on every text change). Shouldn't have to make a ton of objects.