Boiling down CommandParser to no longer do catalog lookups
There's no reason to have the parser take on a dependency to the catalog. The parser should be about parsing, not parsing AND lookup. Removing this dependency simplifies a lot of the workflow in the parser and the corresponding tests. It also means the parser no longer has any dependencies, and can begin to become a static class.

The trade-off here is that the command lookup and null check are now the caller's responsibility--the view model so far. There's probably a cleaner answer here, so we'll see what emerges.