Preview text box is now directly behind the command text
We're beginning to change the idea of preview; instead of parsing the remaining text, we'll put the whole text on a secondary control behind it. This way, the real text obscures it, and we don't have to do any special parsing. This should simplify things a lot and require fewer moving parts.