InputController now ingests an initial input frame
It internally uses a stack to track these, and this counts as the first one--something that we need to guide input from the very start. Externally it only exposes the current frame. We may not need this in the future, but we'll use it for now to get the API going.