Ditching InputText property from ViewModel in favor of a command pattern
The text box changes now are routed as a command to the view model who can respond to it, instead of auto-setting a bound property with all sorts of side-effects and INPC notifications.

Re-worked all the tests and ended up with fewer tests that are shorter, easier to read, with the same coverage.

By splitting the InputText idea into a change action (the command) and a text change action (the event), we're free to let them vary easily, instead of having them tightly coupled together. This was hard to deal with and greatly complicated the tests.