Simplifying the OnCompleteSuggestion command
Generalizing the method that returns whether there's a selected suggestion. Instead of this being a "Can execute" for a specific command, it's now a general-purpose checker that is now reused on the Complete Suggestion path. Was able to ditch the null check from the command handler, since this new CanExecute usage ensures it's only called when it's non-null.